pyproject.toml
[project]
name = "resume-agent"
version = "0.1.0"
description = "Autonomous agent that searches, tailors resumes, and applies to jobs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.20",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "pdfplumber>=0.11.9",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jobspy>=1.1.82",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.48",
    "typer>=0.24.1",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]

[project.scripts]
resume-agent = "resume_agent.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]


[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
]
